requirements.txt
flask
python-dotenv
pygithub
openai
transformers
torch
flask-session
regex